test 요금 정책 단위 테스트 개선
